Refactor code completion logic into dedicated package
Extracted completion-related classes (CompletionItem, CompletionProvider, CompletionResult, JdtCompletionProvider) into a new package for better modularity and reuse. Updated JavaCodeEditorPane and JdtSignatureHelpProvider to use the new completion API, removing duplicated logic and improving maintainability.
