+++ src/main/java/dev/railroadide/railroad/ide/ui/JavaCodeEditorPane.java
@@ -2,6 +2,10 @@ package dev.railroadide.railroad.ide.ui;
 
 import dev.railroadide.core.ui.RRListView;
 import dev.railroadide.railroad.Railroad;
+import dev.railroadide.railroad.ide.completion.CompletionItem;
+import dev.railroadide.railroad.ide.completion.CompletionProvider;
+import dev.railroadide.railroad.ide.completion.CompletionResult;
+import dev.railroadide.railroad.ide.completion.JdtCompletionProvider;
 import dev.railroadide.railroad.ide.signature.JdtSignatureHelpProvider;
 import dev.railroadide.railroad.ide.signature.SignatureHelp;
 import dev.railroadide.railroad.ide.signature.SignatureHelp.ParameterInfo;
@@ -33,7 +37,9 @@ import javafx.scene.text.TextFlow;
 import javafx.stage.Popup;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.core.compiler.IProblem;
-import org.eclipse.jdt.core.dom.*;
+import org.eclipse.jdt.core.dom.AST;
+import org.eclipse.jdt.core.dom.ASTParser;
+import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.fxmisc.richtext.event.MouseOverTextEvent;
 import org.fxmisc.richtext.model.PlainTextChange;
 import org.fxmisc.richtext.model.StyleSpans;
@@ -85,6 +91,7 @@ public class JavaCodeEditorPane extends TextEditorPane {
     private final List<CompletionItem> completionCandidates = new ArrayList<>();
     private volatile int completionDotIndex = -1;
 
+    private final CompletionProvider completionProvider;
     private final Popup signaturePopup = new Popup();
     private final TextFlow signatureTextFlow = new TextFlow();
     private final AtomicInteger signatureGeneration = new AtomicInteger();
@@ -96,6 +103,7 @@ public class JavaCodeEditorPane extends TextEditorPane {
     public JavaCodeEditorPane(Project project, Path item) {
         super(item);
         this.project = Objects.requireNonNull(project, "project");
+        this.completionProvider = new JdtCompletionProvider(filePath, SYSTEM_MODULE_PATHS);
         this.signatureHelpProvider = new JdtSignatureHelpProvider(filePath, SYSTEM_MODULE_PATHS);
 
         diagnosticPopup.setAutoHide(true);
@@ -452,9 +460,8 @@ public class JavaCodeEditorPane extends TextEditorPane {
             return null;
 
         for (ProblemDiagnostic diagnostic : visibleDiagnostics) {
-            if (index >= diagnostic.start() && index <= diagnostic.end()) {
+            if (index >= diagnostic.start() && index <= diagnostic.end())
                 return diagnostic;
-            }
         }
 
         return null;
@@ -492,9 +499,15 @@ public class JavaCodeEditorPane extends TextEditorPane {
     private void triggerCompletion(int dotIndex) {
         String snapshot = getText();
         int generation = completionGeneration.incrementAndGet();
-        CompletableFuture.supplyAsync(() -> computeCompletion(snapshot, dotIndex), worker)
+        CompletableFuture
+            .supplyAsync(() -> completionProvider.compute(snapshot, dotIndex), worker)
             .thenAccept(result -> Platform.runLater(() -> {
-                if (completionGeneration.get() == generation) {
+                if (completionGeneration.get() != generation)
+                    return;
+
+                if (result == null || result.items().isEmpty()) {
+                    hideAutoComplete();
+                } else {
                     handleCompletionResult(result);
                 }
             }))
@@ -502,277 +515,6 @@ public class JavaCodeEditorPane extends TextEditorPane {
                 Railroad.LOGGER.error("Failed to compute Java code completion", throwable);
                 return null;
             });
-    }
-
-    private CompletionResult computeCompletion(String text, int dotIndex) {
-        if (dotIndex < 0 || dotIndex >= text.length() || text.charAt(dotIndex) != '.')
-            return CompletionResult.empty();
-
-        Pair range = findIdentifierRangeBeforeDot(text, dotIndex);
-        if (range == null)
-            return CompletionResult.empty();
-
-        ASTParser parser = ASTParser.newParser(AST.JLS21); // TODO: Detect version from project settings
-        parser.setSource(text.toCharArray());
-        parser.setKind(ASTParser.K_COMPILATION_UNIT);
-        parser.setResolveBindings(true);
-        parser.setBindingsRecovery(true);
-        parser.setStatementsRecovery(true);
-        parser.setUnitName(filePath.getFileName().toString());
-
-        // TODO: Set classpath and sourcepath from project settings
-        parser.setEnvironment(SYSTEM_MODULE_PATHS, null, null, false);
-
-        CompilationUnit unit = (CompilationUnit) parser.createAST(null);
-        var finder = new NodeFinder(unit, range.start(), Math.max(1, range.length()));
-        ASTNode node = finder.getCoveredNode();
-        if (node == null) {
-            node = finder.getCoveringNode();
-        }
-
-        CompletionTarget target = null;
-        ASTNode cursor = node;
-        while (cursor != null && target == null) {
-            if (cursor instanceof ExpressionStatement statement) {
-                cursor = statement.getExpression();
-            }
-
-            target = resolveCompletionTarget(cursor);
-            cursor = cursor != null ? cursor.getParent() : null;
-        }
-
-        if (target == null || target.type() == null) {
-            return CompletionResult.empty();
-        }
-
-        List<CompletionItem> suggestions = collectCompletionSuggestions(unit, target);
-        if (suggestions.isEmpty())
-            return CompletionResult.empty();
-
-        return new CompletionResult(dotIndex, suggestions);
-    }
-
-    private CompletionTarget resolveCompletionTarget(ASTNode node) {
-        if (node == null)
-            return null;
-
-        if (node instanceof Name name) {
-            IBinding binding = name.resolveBinding();
-            if (binding instanceof ITypeBinding typeBinding)
-                return new CompletionTarget(normalizeType(typeBinding), true);
-
-            if (binding instanceof IVariableBinding variableBinding) {
-                ITypeBinding type = normalizeType(variableBinding.getType());
-                if (type != null)
-                    return new CompletionTarget(type, false);
-            }
-
-            if (binding instanceof IMethodBinding methodBinding) {
-                ITypeBinding type = normalizeType(methodBinding.getReturnType());
-                if (type != null)
-                    return new CompletionTarget(type, false);
-            }
-        }
-
-        if (node instanceof Expression expression) {
-            ITypeBinding type = normalizeType(expression.resolveTypeBinding());
-            if (type != null)
-                return new CompletionTarget(type, false);
-        }
-
-        return null;
-    }
-
-    private ITypeBinding normalizeType(ITypeBinding binding) {
-        if (binding == null)
-            return null;
-
-        if (binding.isAnonymous()) {
-            binding = binding.getSuperclass();
-        }
-
-        if (binding == null)
-            return null;
-
-        if (binding.isArray() || binding.isPrimitive() || binding.isNullType())
-            return binding;
-
-        if (binding.isTypeVariable()) {
-            ITypeBinding erasure = binding.getErasure();
-            return erasure != null ? erasure : binding;
-        }
-
-        ITypeBinding declaration = binding.getTypeDeclaration();
-        return declaration != null ? declaration : binding;
-    }
-
-    private List<CompletionItem> collectCompletionSuggestions(CompilationUnit unit, CompletionTarget target) {
-        LinkedHashSet<CompletionItem> results = new LinkedHashSet<>();
-        String currentPackage = extractPackageName(unit);
-        Set<String> visited = new HashSet<>();
-        addMembers(target.type(), target.staticContext(), currentPackage, results, visited);
-
-        return new ArrayList<>(results);
-    }
-
-    private String extractPackageName(CompilationUnit unit) {
-        if (unit == null)
-            return "";
-
-        PackageDeclaration declaration = unit.getPackage();
-        if (declaration == null || declaration.getName() == null)
-            return "";
-
-        return declaration.getName().getFullyQualifiedName();
-    }
-
-    private void addMembers(ITypeBinding type, boolean staticContext, String currentPackage,
-                            LinkedHashSet<CompletionItem> results, Set<String> visited) {
-        if (type == null)
-            return;
-
-        ITypeBinding working = type;
-        if (!working.isArray() && !working.isPrimitive()) {
-            working = working.getErasure();
-        }
-
-        String key = working.getKey();
-        if (key == null) {
-            key = working.getQualifiedName();
-        }
-
-        if (key == null || !visited.add(key))
-            return;
-
-        if (working.isArray()) {
-            results.add(new CompletionItem("length", "length : int"));
-            addMembers(working.getSuperclass(), staticContext, currentPackage, results, visited);
-            return;
-        }
-
-        if (working.isPrimitive())
-            return;
-
-        for (IVariableBinding field : working.getDeclaredFields()) {
-            if (field.isSynthetic())
-                continue;
-
-            int modifiers = field.getModifiers();
-            boolean isStatic = Modifier.isStatic(modifiers);
-            if (staticContext && !isStatic)
-                continue;
-            else if (isStatic)
-                continue;
-
-            if (!isAccessible(modifiers, field.getDeclaringClass(), currentPackage))
-                continue;
-
-            results.add(createFieldItem(field));
-        }
-
-        for (IMethodBinding method : working.getDeclaredMethods()) {
-            if (method.isConstructor() || method.isSynthetic())
-                continue;
-
-            int modifiers = method.getModifiers();
-            boolean isStatic = Modifier.isStatic(modifiers);
-            if (staticContext && !isStatic)
-                continue;
-            else if (isStatic)
-                continue;
-
-            if (!isAccessible(modifiers, method.getDeclaringClass(), currentPackage))
-                continue;
-
-            results.add(createMethodItem(method));
-        }
-
-        if (staticContext) {
-            for (ITypeBinding nested : working.getDeclaredTypes()) {
-                if (nested == null || nested.getName() == null || nested.getName().isBlank())
-                    continue;
-
-                int modifiers = nested.getModifiers();
-                if (!Modifier.isStatic(modifiers))
-                    continue;
-
-                if (!isAccessible(modifiers, nested, currentPackage))
-                    continue;
-
-                results.add(createTypeItem(nested));
-            }
-        }
-
-        addMembers(working.getSuperclass(), staticContext, currentPackage, results, visited);
-        for (ITypeBinding iface : working.getInterfaces()) {
-            addMembers(iface, staticContext, currentPackage, results, visited);
-        }
-    }
-
-    private CompletionItem createFieldItem(IVariableBinding field) {
-        String name = field.getName();
-        String type = renderType(field.getType());
-        String prefix = Modifier.isStatic(field.getModifiers()) ? "static " : "";
-        return new CompletionItem(name, prefix + name + " : " + type);
-    }
-
-    private CompletionItem createMethodItem(IMethodBinding method) {
-        String name = method.getName();
-        String prefix = Modifier.isStatic(method.getModifiers()) ? "static " : "";
-        String parameters = Arrays.stream(method.getParameterTypes())
-            .map(this::renderType)
-            .collect(Collectors.joining(", "));
-        String returnType = renderType(method.getReturnType());
-        String display = prefix + name + "(" + parameters + ") : " + returnType;
-        return new CompletionItem(name, display);
-    }
-
-    private CompletionItem createTypeItem(ITypeBinding type) {
-        String name = type.getName();
-        String prefix = Modifier.isStatic(type.getModifiers()) ? "static " : "";
-        return new CompletionItem(name, prefix + name + " (type)");
-    }
-
-    private String renderType(ITypeBinding type) {
-        if (type == null)
-            return "?";
-
-        if (type.isArray())
-            return renderType(type.getComponentType()) + "[]";
-
-        if (type.isPrimitive() || type.isNullType())
-            return type.getName();
-
-        ITypeBinding erasure = type.getErasure();
-        String qualified = erasure != null ? erasure.getQualifiedName() : type.getQualifiedName();
-        if (qualified != null && !qualified.isBlank())
-            return qualified;
-
-        String name = type.getName();
-        return name != null && !name.isBlank() ? name : "?";
-    }
-
-    private boolean isAccessible(int modifiers, ITypeBinding declaringType, String currentPackage) {
-        if (Modifier.isPublic(modifiers) || Modifier.isProtected(modifiers))
-            return true;
-
-        if (Modifier.isPrivate(modifiers))
-            return false;
-
-        String declaringPackage = declaringType != null ? packageName(declaringType) : "";
-        return Objects.equals(declaringPackage, currentPackage);
-    }
-
-    private String packageName(ITypeBinding type) {
-        if (type == null)
-            return "";
-
-        IPackageBinding pkg = type.getPackage();
-        if (pkg == null)
-            return "";
-
-        String name = pkg.getName();
-        return name != null ? name : "";
     }
 
     private void handleCompletionResult(CompletionResult result) {
@@ -1017,19 +759,6 @@ public class JavaCodeEditorPane extends TextEditorPane {
         return column;
     }
 
-    private static Pair findIdentifierRangeBeforeDot(String text, int dotPosition) {
-        int start = dotPosition - 1;
-        while (start >= 0 && Character.isJavaIdentifierPart(text.charAt(start))) {
-            start--;
-        }
-
-        start++;
-        if (start > dotPosition)
-            return null;
-
-        return new Pair(start, dotPosition - start);
-    }
-
     public String getLanguageId() {
         return "java";
     }
@@ -1119,25 +848,6 @@ public class JavaCodeEditorPane extends TextEditorPane {
         public String getMessage(Locale locale) {
             return message;
         }
-    }
-
-    private record CompletionTarget(ITypeBinding type, boolean staticContext) {
-    }
-
-    private record CompletionResult(int dotIndex, List<CompletionItem> items) {
-        static CompletionResult empty() {
-            return new CompletionResult(-1, Collections.emptyList());
-        }
-    }
-
-    private record CompletionItem(String insertText, String displayText) {
-        @Override
-        public @NotNull String toString() {
-            return displayText;
-        }
-    }
-
-    private record Pair(int start, int length) {
     }
 
     private static final class CompletionItemListCell extends ListCell<CompletionItem> {
